+++ ktor-client-utils/src/nativeTest/kotlin/com/pambrose/common/dsl/DefaultEngine.native.kt
@@ -1,0 +1,3 @@
+package com.pambrose.common.dsl
+
+internal actual val hasDefaultEngine: Boolean = false